ovslatency: Initial support for measuring OVS latency
Uses kprobe on entry and return of ovs_vport_receive (seems to
be exported across kernel versions) to generate a histogram of
the latency added by OVS processing.

Right now it only works with the legacy kprobe setup. Example:
    M=/lib/modules/$(uname -r)/kernel/net/openvswitch/openvswitch.ko
    sudo perf probe -m $M ovs_vport_receive
    sudo perf probe -m $M ovs_vport_receive_ret=ovs_vport_receive%return
    sudo src/obj/ovslatency

    $ sudo bin/ovslatency -f bin/ovslatency.o
    02:01:07.088235: total number of packets 429576:
        time (usec)        count
           0  -   10:     354767
         10+  -   25:      26936
         25+  -   50:      25406
         50+  -  100:      15970
        100+  -  250:       5447
        250+  -  500:        528
        500+  -   up:        522

Plan to add support for newer kernels and many ideas on followups.

Signed-off-by: David Ahern <dsahern@gmail.com>