pktlatency: Parameterize buckets
Use macro for defining the buckets for packet latency to make
it easier to change the values. Macros are in usec, so update
latency_gen_sample.

Change bpf program to invert the checking order so that fastest
packets have the fewest checks.

Add new bucket for running sum to check the average packet
latency.

Signed-off-by: David Ahern <dsahern@gmail.com>