cleanup kprobe handling
kprobes are now created on the fly using old API for kernels that can
not create a probe using perf_event. Update the kprobe management
infra to take struct kprobe_data that defines the program name, function
name, whether it is a return probe and save the fd for that probe.
Create init and cleanup functions and make sure the cleanup is done
properly at exit of all commands relying on kprobes.

Signed-off-by: David Ahern <dsahern@gmail.com>